--static in build instructions should be --static-dir
Also, setup.py did not check if 'gmp' was in the list of libraries
before trying to remove it, leading to ValueError possibly being raised.
It was also necessary to change 'gmp' to 'mpir' in the list of
libraries. This may be because this list is used before
build_extensions() has a chance to modify it.